Use static_cast to reduce compiler warnings.

+++ snappy-test.h
@@ -143,7 +143,7 @@ class CycleTimer {
 
     double elapsed = static_cast<double>(stop.QuadPart - start_.QuadPart) /
         frequency.QuadPart;
-    real_time_us_ += elapsed * 1e6 + 0.5;
+    real_time_us_ += static_cast<int64_t>(elapsed * 1e6 + 0.5);
 #else
     struct ::timeval stop;
     ::gettimeofday(&stop, nullptr);
